(BC) Separation of AbstractRepository into Trait and Interface (#1269)
